Modularize the CPUID code into cpuid.c/h
