+++ test-sgx.h
@@ -1,0 +1,16 @@
+///////////////////////////////////////////////////////////////////////////////
+//  test-sgx.h - 2023
+//
+/// Things that every module in test-sgx needs to know.
+///
+/// @file   test-sgx.h
+/// @author Lars Luhr   <mail@ayeks.de>
+/// @author Mark Nelson <marknels@hawaii.edu>
+///////////////////////////////////////////////////////////////////////////////
+#pragma once
+
+#include <stdlib.h>    // For exit() EXIT_FAILURE
+
+#define PROGRAM_NAME "test-sgx"
+
+#define EXIT_ON_FAILURE // exit( EXIT_FAILURE );
